requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pymongo==4.12.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-surprise==1.1.4
scipy==1.15.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3